feat: Improve main page with movie finder title and background color
This commit improves the main page by adding a movie finder title and changing the background color to black. It also includes the necessary HTML and CSS changes to achieve this. The movie finder title is displayed as a heading with a blue color, and the background color is set to black using the `bg-black` class. These changes enhance the visual appearance and branding of the application.