feat: Update viewDocument function to accept pageHref parameter
This commit updates the viewDocument function in Main.roc to accept a pageHref parameter. This allows for dynamic loading of different pages based on the provided URL. The function is modified to use the pageHref parameter instead of the hardcoded "/home" value when making an HTTP GET request. This change improves the flexibility and reusability of the viewDocument function.